+++ src/services/nest.rs
@@ -1,4 +1,4 @@
-use crate::message::{Message, Reading, Type};
+use crate::message::{Message, Type};
 use crate::supervisor;
 use crate::value::Value;
 use crate::Result;
@@ -48,37 +48,31 @@ fn send_readings(service_id: &str, event: &NestEvent, tx: &Sender<Message>) -> R
     let now = Local::now();
 
     for (id, thermostat) in event.data.devices.thermostats.iter() {
-        tx.send(Message {
-            type_: Type::Actual,
-            reading: Reading {
-                sensor: format!("{}::thermostat::{}::ambient_temperature", service_id, &id),
-                value: Value::Celsius(thermostat.ambient_temperature_c),
-                timestamp: now,
-            },
-        })?;
-        tx.send(Message {
-            type_: Type::Actual,
-            reading: Reading {
-                sensor: format!("{}::thermostat::{}::humidity", service_id, &id),
-                value: Value::Rh(thermostat.humidity),
-                timestamp: now,
-            },
-        })?;
+        tx.send(Message::new(
+            Type::ReadLogged,
+            format!("{}::thermostat::{}::ambient_temperature", service_id, &id),
+            Value::Celsius(thermostat.ambient_temperature_c),
+            now,
+        ))?;
+        tx.send(Message::new(
+            Type::ReadLogged,
+            format!("{}::thermostat::{}::humidity", service_id, &id),
+            Value::Rh(thermostat.humidity),
+            now,
+        ))?;
     }
 
     for (id, camera) in event.data.devices.cameras.iter() {
-        tx.send(Message {
-            type_: Type::Actual,
-            reading: Reading {
-                sensor: format!("{}::camera::{}::snapshot_url", service_id, &id),
-                value: Value::ImageUrl(camera.snapshot_url.clone()),
-                timestamp: now,
-            },
-        })?;
+        tx.send(Message::new(
+            Type::ReadLogged,
+            format!("{}::camera::{}::snapshot_url", service_id, &id),
+            Value::ImageUrl(camera.snapshot_url.clone()),
+            now,
+        ))?;
 
         if let Some(ref event) = camera.last_event {
             tx.send(Message::new(
-                Type::Actual,
+                Type::ReadLogged,
                 format!("{}::camera::{}::animated_image_url", service_id, &id),
                 Value::ImageUrl(event.animated_image_url.clone()),
                 event.start_time,